[FIX] sequence number generation
To set a sequence number in Psql after the table is filled, we get the
highest number in use, add one and set the sequence to this. The
existing code retrieved the highest number in use using
$max = max_rs(). This actually counts the elements in the set and thus
is wrong when holes exist. The correct function is max() and I use this
one now.