[TEST] Windows CLI testing now TODO
CLI testing does not work as expected on Windows. Therefore, I marked
that as SKIP before. Actually, its better if DBICx::Backend::Move one
day fully works on Win too, including this test. Therefore, it is a TODO
and the test should show this. Thanks to ilmari on #perl-help for
pointing that out to me.